Split out filesystem system calls in the libc backend.
